requirements.txt
# Core FastAPI and server dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0

# AI and LLM dependencies
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0

# OCR and image processing
opencv-python>=4.8.0
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.1.0
numpy>=1.24.0

# PDF processing
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
reportlab>=4.0.0

# Audio and TTS
gTTS>=2.4.0
pydub>=0.25.0

# Data processing and utilities
pandas>=2.0.0
requests>=2.31.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

fastapi
uvicorn
python-multipart
pydantic
easyocr
opencv-python-headless
pymupdf
PyPDF2
openai
gtts
